+++ eva/src/components/dashBoardInside/dashConstructorScheme/dashConstructorSchemes.vue
@@ -686,6 +686,7 @@ export default {
     return {
       actions: [
         { name: 'click:label', capture: ['value1', 'value2', 'value3', 'value4', 'value5'] },
+        { name: 'click:image', capture: [] },
       ],
       isEdit: false,
       gear: mdiSettings,
@@ -793,6 +794,19 @@ export default {
         return savedGraph[this.localActiveSchemeId] || [];
       }
       return [];
+    },
+    captureFromImageNode() {
+      const result = [];
+      const prepareItems = [];
+      this.savedGraphObject
+        .filter((item) => item.data.tag.dataType === 'image-node' && item.data.tag.fromOtl)
+        .forEach((item) => {
+          prepareItems.push(...Object.keys(item.data.tag.fromOtl));
+        });
+      new Set(prepareItems).forEach((item) => {
+        result.push(item);
+      });
+      return result;
     },
     innerSize() {
       return {
@@ -905,6 +919,31 @@ export default {
     loadingSearchForBuildScheme(val) {
       this.toggleLoading(val);
     },
+    dataSelectedNode(node) {
+      if (node?.dataType === 'image-node') {
+        this.actions = this.actions.map((action) => {
+          if (action.name !== 'click:image') {
+            return action;
+          }
+          return {
+            ...action,
+            capture: Object.keys(node),
+          };
+        });
+      }
+    },
+    captureFromImageNode(imageCapture) {
+      this.actions = this.actions.map((action) => {
+        if (action.name === 'click:label') {
+          return action;
+        }
+        return {
+          ...action,
+          capture: imageCapture,
+        };
+      });
+      this.setActions();
+    },
   },
   mounted() {
     this.createGraph();
@@ -913,6 +952,17 @@ export default {
     this.setActions();
     this.localActiveSchemeId = this.activeSchemeId;
     this.isEdit = this.dashboardEditMode;
+    if (this.constructorSchemes) {
+      if (this.isAlwaysUpdateScheme && this.dataForBuildScheme?.length > 0) {
+        this.constructorSchemes.buildSchemeFromSearch(
+          this.dataForBuildScheme,
+          this.minimumLastSegmentLength,
+          this.minimumEdgeToEdgeDistance,
+        );
+      } else {
+        this.constructorSchemes.update(this.savedGraphObject);
+      }
+    }
   },
   methods: {
     setActions() {
@@ -975,7 +1025,7 @@ export default {
         isEdit: this.dashboardEditMode,
         isBridgesEnable: this.isBridgeEnable,
         onClickObject: (type, data) => {
-          if (type !== 'label-type-0') {
+          if (!type.includes('label-type') && type !== 'image-node') {
             return;
           }
           const { tockens: tokens } = this.$store.state[this.idDashFrom];
@@ -986,11 +1036,21 @@ export default {
               capture,
               elem,
             }) => {
-              if (elem === this.idFrom && action === 'click:label' && data[capture]) {
+              if (
+                elem === this.idFrom
+                  && ((action === 'click:label' && data[capture])
+                      || (action === 'click:image'))
+              ) {
+                let value;
+                if (action === 'click:image') {
+                  value = data.fromOtl[capture] || 'Нет данных';
+                } else {
+                  value = data[capture];
+                }
                 this.$store.commit('setTocken', {
                   token: { name, action, capture },
                   idDash: this.idDashFrom,
-                  value: data[capture],
+                  value,
                   store: this.$store,
                 });
               }