+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Constructor-schemes: the ability to remove a scheme from the list of saved schemes
+- Constructor-schemes: token-action click:image
 
 ### Changed